Move all ffi persister adapters to persist.rs
Move the preexisting JSON persister traits and callback adapters to
persist.rs for clearer organization.
